First character in a string that occurs twice
This problem has been solved using HashMap Collection data type. The character and its frequency is stored in the HasMap. If a character occurs more than once its frequency is updated accordingly. Then another loop goes through the map and checks for the character that has frequency=2. When the first character with frequency 2 is found, the loop prints result and terminates. If no character occurs twice in the string, an appropriate message is printed.